wip: authors
